requirements.txt
python-telegram-bot[job-queue]>=21.0,<22
aiosqlite>=0.19.0,<1
openai>=1.12.0,<2
loguru>=0.7.2,<1
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
python-dotenv>=1.0,<2
python-dateutil>=2.8,<3
pytz>=2024.1
pytest>=8.0,<9
pytest-asyncio>=0.23,<1
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
python-jose[cryptography]>=3.3,<4
httpx>=0.27,<1
honcho>=1.1,<2
